refactor(google): extract gke utils as gke_utils module
